refactor: move group_receipts_by_shard and build_receipts_hashes to c… (#6248)
…hain
