+++ chain/chunks/src/test_utils.rs
@@ -4,7 +4,7 @@ use near_primitives::time::Clock;
 
 use near_chain::test_utils::KeyValueRuntime;
 use near_chain::types::{RuntimeAdapter, Tip};
-use near_chain::ChainStore;
+use near_chain::{Chain, ChainStore};
 use near_crypto::KeyType;
 use near_network::test_utils::MockPeerManagerAdapter;
 use near_primitives::block::BlockHeader;
@@ -218,7 +218,7 @@ impl ChunkTestFixture {
         );
         let receipts = Vec::new();
         let shard_layout = mock_runtime.get_shard_layout(&EpochId::default()).unwrap();
-        let receipts_hashes = mock_runtime.build_receipts_hashes(&receipts, &shard_layout);
+        let receipts_hashes = Chain::build_receipts_hashes(&receipts, &shard_layout);
         let (receipts_root, _) = merkle::merklize(&receipts_hashes);
         let (mock_chunk, mock_merkles) = producer_shard_manager
             .create_encoded_shard_chunk(
