+++ integration-tests/src/tests/client/challenges.rs
@@ -7,7 +7,8 @@ use near_chain::missing_chunks::MissingChunksPool;
 use near_chain::types::BlockEconomicsConfig;
 use near_chain::validate::validate_challenge;
 use near_chain::{
-    Block, ChainGenesis, ChainStoreAccess, DoomslugThresholdMode, Error, ErrorKind, Provenance,
+    Block, Chain, ChainGenesis, ChainStoreAccess, DoomslugThresholdMode, Error, ErrorKind,
+    Provenance,
 };
 use near_chain_configs::Genesis;
 use near_client::test_utils::{create_chunk, create_chunk_with_transactions, run_catchup, TestEnv};
@@ -475,11 +476,9 @@ fn test_receive_invalid_chunk_as_chunk_producer() {
     // But everyone who doesn't track this shard have accepted.
     let shard_layout =
         env.clients[0].runtime_adapter.get_shard_layout(&EpochId::default()).unwrap();
-    let receipts_hashes =
-        env.clients[0].runtime_adapter.build_receipts_hashes(&receipts, &shard_layout);
+    let receipts_hashes = Chain::build_receipts_hashes(&receipts, &shard_layout);
     let (_receipts_root, receipts_proofs) = merklize(&receipts_hashes);
-    let receipts_by_shard =
-        env.clients[0].shards_mgr.group_receipts_by_shard(receipts, &shard_layout);
+    let receipts_by_shard = Chain::group_receipts_by_shard(receipts, &shard_layout);
     let one_part_receipt_proofs = env.clients[0].shards_mgr.receipts_recipient_filter(
         0,
         Vec::default(),
